+++ packages/play/src/main.ts
@@ -6,6 +6,7 @@ import router from '@/router'
 import PlusProComponents from '../../plus-pro-components/index'
 import App from './App.vue'
 import './style.css'
+import 'element-plus/dist/index.css'
 
 const app = createApp(App)
 app